Add --disable-scrollkeeper and --disable-update-mimedb options to
* Makefile.am:
* configure.in:
* help/omf.make:
Add --disable-scrollkeeper and --disable-update-mimedb options to
configure to skip building the scrollkeeper and mime database on
make install. The files generated by this must not be shipped by
binary packages and having this options makes it easier to package
genius.

svn path=/trunk/; revision=589

+++ help/omf.make
@@ -35,27 +35,35 @@ scrollkeeper_localstate_dir = $(localstatedir)/scrollkeeper
 omf: omf_timestamp
 
 omf_timestamp: $(omffile)
+if ENABLE_SK
 	-for file in $(omffile); do \
 	  scrollkeeper-preinstall $(docdir)/$(docname).xml $(srcdir)/$$file $$file.out; \
 	done; \
 	touch omf_timestamp
+endif
 
 install-data-hook-omf:
+if ENABLE_SK
 	$(mkinstalldirs) $(DESTDIR)$(omf_dest_dir)
 	for file in $(omffile); do \
 		$(INSTALL_DATA) $$file.out $(DESTDIR)$(omf_dest_dir)/$$file; \
 	done
 	-scrollkeeper-update -p $(DESTDIR)$(scrollkeeper_localstate_dir) -o $(DESTDIR)$(omf_dest_dir)
+endif
 
 uninstall-local-omf:
+if ENABLE_SK
 	-for file in $(srcdir)/*.omf; do \
 		basefile=`basename $$file`; \
 		rm -f $(DESTDIR)$(omf_dest_dir)/$$basefile; \
 	done
 	-rmdir $(DESTDIR)$(omf_dest_dir)
 	-scrollkeeper-update -p $(DESTDIR)$(scrollkeeper_localstate_dir)
+endif
 
 clean-local-omf:
+if ENABLE_SK
 	-for file in $(omffile); do \
 		rm -f $$file.out; \
 	done
+endif
